requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.5.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.2
django-rest-framework-mongoengine==3.4.1
django_celery_results==2.6.0
djangorestframework==3.16.0
dnspython==2.7.0
gunicorn==23.0.0
jalali_core==1.0.0
jdatetime==5.2.0
kombu==5.5.3
mongoengine==0.29.1
packaging==24.2
prompt_toolkit==3.0.51
pymongo==4.12.0
python-dateutil==2.9.0.post0
redis==5.2.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
